+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - Extracted launching functions from `Terminus` (base class) and moved them into new `LaunchHelper` class. (#882)
 - Extracted template function from `Utils` and moved it into new `TemplateHelper` class. (#884)
 - Changed the site-DNE message to instruct the user to run `sites list`. (#895)
+- `site set-service-level` now also accepts levels of "professional", "sandbox", and "personal". (#894)
 
 ### Removed
 - `addHostnames`, `deleteHostnames()`, and `getHostnames()` has been removed from `Environment`. Use new hostnames property (contains Hostnames collection) instead. (#860)